separate logout from firebase helper
